refactor: Extract and prepare for bar chart
